- Entities now convert properly for PS4 - Fixed rare chunk issue - LevelFile now converts "generatorOptions" - Fixed picture.cpp incorrectly stretching images - Fixed detectConsole.hpp (can now detect between all LCE savefile variants) - Removed forced file removal from conversion, preventing removal of files that usually cause issues is reserved to converting with the config. - Updated License to GNU GPL v3. - More options for "conversion.json"

+++ code/Impl/levelFile.hpp
@@ -18,7 +18,7 @@ namespace editor {
             adventure
         };
 
-        enum class Difficulty : i32 {
+        enum class Difficulty : u8 {
             peaceful,
             easy,
             medium,
@@ -69,7 +69,8 @@ namespace editor {
         /* TU14 */ std::optional<u8> m_initialized;
         /* TU17 */ std::optional<i32> m_HellScale;
         /* TU19 */ std::optional<i64> m_DayTime;
-        /*19-23 */ std::optional<std::string> m_generatorOptions;
+        // TODO: used to be "std::optional<std::string>" (?) but latest versions use a std::vector<u8>
+        /*19-23 */ std::optional<NBTByteArray> m_generatorOptions;
         /* TU25 */ std::optional<enums::Difficulty> m_Difficulty;
         /* TU25 */ std::optional<i32> m_clearWeatherTime;
         /* TU31 */ std::optional<u8> m_DifficultyLocked;
